requirements.txt
socket
requests
socket
dns.resolver
beautifulsoup4
